+++ main.go
@@ -81,14 +81,14 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "Cluster")
 		os.Exit(1)
 	}
-	//if err = (&controllers.CAPAClusterReconciler{
-	//	Client: mgr.GetClient(),
-	//	Log:    ctrl.Log.WithName("capa-controller"),
-	//	Scheme: mgr.GetScheme(),
-	//}).SetupWithManager(mgr); err != nil {
-	//	setupLog.Error(err, "unable to create controller", "controller", "Cluster")
-	//	os.Exit(1)
-	//}
+	if err = (&controllers.CAPAClusterReconciler{
+		Client: mgr.GetClient(),
+		Log:    ctrl.Log.WithName("capa-controller"),
+		Scheme: mgr.GetScheme(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Cluster")
+		os.Exit(1)
+	}
 	//+kubebuilder:scaffold:builder
 
 	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {